Add --namespace-labels, --node-labels, and --pod-labels flags
This adds new flags to kube-capacitor allowing to filter metrics by labels on
namespaces, nodes, and pods.

Any set of labels can be provided:

```
kube-capacity -p -u --namespace-labels "mongodb=true" --pod-labels "app=mongodb"
```

Labels are parsed the same way that kubectl parses them.

The cluster metrics are all fixed to be the sum of the filtered resources, so the cluster and node metrics
do not include metrics from pods that do not fit into the namespace, node or pod labels.