Add a flag to exclude tainted nodes
On a cluster that has many tainted nodes it's not straightforward to understand
whether there is sufficient capacity to drain one: nodes with taints can make
it seem like there is more capacity than there really would be.

This patch solves for the simplest use case by allowing all tainted nodes to be
excluded with a flag.